requirements.txt
tkinter==8.6.13
pillow==11.2.1
sqlite3==3.50.0